requirements.txt
Scrapy==2.5.1
# lxml==4.6.3
# pyOpenSSL==21.0.0
# cryptography==3.4.7
# cssselect==1.1.0
pymongo==3.12.0
python-dotenv
pandas==1.5.3